Moved distance_normalmaps to engine/ since it doesn't depend on meshers
