+++ terrain/variable_lod/voxel_lod_terrain.cpp
@@ -1,11 +1,11 @@
 #include "voxel_lod_terrain.h"
 #include "../../constants/voxel_string_names.h"
 #include "../../edition/voxel_tool_lod_terrain.h"
+#include "../../engine/distance_normalmaps.h"
 #include "../../engine/load_all_blocks_data_task.h"
 #include "../../engine/voxel_engine_gd.h"
 #include "../../engine/voxel_engine_updater.h"
 #include "../../meshers/blocky/voxel_mesher_blocky.h"
-#include "../../meshers/transvoxel/distance_normalmaps.h"
 #include "../../meshers/transvoxel/voxel_mesher_transvoxel.h"
 #include "../../storage/voxel_buffer_gd.h"
 #include "../../util/container_funcs.h"
